+++ app/(auth)/callback/page.tsx
@@ -1,4 +1,4 @@
-import AuthCallbackContainer from '@/components/AuthCallback/AuthCallbackContainer';
+import AuthCallbackContainer from '@/components/auth/AuthCallback/AuthCallbackContainer';
 
 // 구글에서 code를 받아온 뒤 리다이렉트되는 페이지
 export default function AuthCallbackPage() {
