+++ components/articles/DeleteButton/DeleteButtonContainer.tsx
@@ -5,7 +5,7 @@ import { useRouter } from 'next/navigation';
 import Swal from 'sweetalert2';
 import { getCookie } from 'cookies-next';
 
-import DeleteButtonView from '@/components/DeleteButton/DeleteButtonView';
+import DeleteButtonView from '@/components/articles/DeleteButton/DeleteButtonView';
 import axiosInstance from '@/utils/axios';
 
 interface PropsType {
@@ -16,9 +16,10 @@ interface PropsType {
 export default function DeleteButtonContainer({ articleId, authorNickname }: PropsType) {
   const router = useRouter();
   const [nickname, setNickname] = useState('');
-  const accessToken = getCookie('accessToken');
 
   useEffect(() => {
+    const accessToken = getCookie('accessToken');
+
     async function getUserNickname() {
       // 비로그인 유저는 닉네임을 가져올 수 없음
       if (!accessToken) return;
