+++ components/articles/DeleteButton/DeleteButtonView.tsx
