+++ components/common/SearchBar/SearchBarContainer.tsx
@@ -1,4 +1,4 @@
-import SearchBarView from '@/components/SearchBar/SearchBarView';
+import SearchBarView from '@/components/common/SearchBar/SearchBarView';
 
 export default function SearchBarContainer() {
   return <SearchBarView />;
