+++ components/common/SearchBar/SearchBarView.tsx
